Simplified pipeline that doesn't use watershed CSF segmentation
This pipeline is intended for cases in which the ventricles
are relatively normal, and the images have thick slices,
i.e. voxels are highly anisotropic. In such cases the
initial SPM csf segmentation is OK, but the watershed
segmentation of csf, intended to correct problem SPM
segmentations, introduces problems as a result of the
thick slices (i.e is likely to incorrectly segment
tissue due to partial volume effects). This problem
tends not to occur with relatively isotropic voxels.

The watershed step is replaced by a simple thresholding.